% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCoefsfromModel}
\alias{getCoefsfromModel}
\title{getCoefsfromModel}
\usage{
getCoefsfromModel(object)
}
\description{
get variables' names from the fitted model objects in different classes
}
\keyword{internal}
