% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surr_rsq.R, R/surr_rsq_ci.R, R/surr_rsq_rank.R
\name{print}
\alias{print}
\alias{print.surr_rsq}
\alias{print.surr_rsq_ci}
\alias{print.surr_rsq_rank}
\title{Print surrogate R-squared measure}
\usage{
\method{print}{surr_rsq}(x, digits = max(2, getOption("digits") - 2), ...)

\method{print}{surr_rsq_ci}(x, digits = max(2, getOption("digits") - 2), ...)

\method{print}{surr_rsq_rank}(x, digits = max(2, getOption("digits") - 2), ...)
}
\arguments{
\item{x}{A surr_rsq_rank object for printing out results.}

\item{digits}{A default number to specify decimal digit values.}

\item{...}{Additional optional arguments.}
}
\value{
Print surrogate R-squared measure of a surr_rsq object

Print surrogate R-squared confidence interval measure

Print surrogate R-squared ranking table from a surr_rsq_rank object
}
\description{
Print surrogate R-squared measure

Print surrogate R-squared confidence interval measure

Print surrogate R-squared ranking table
}
\keyword{internal}
