% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsq.R
\name{rsq}
\alias{rsq}
\title{R2}
\usage{
rsq(
  model,
  full_model = NULL,
  data,
  which = c("Surrogate", "McFadden", "McKelveyZavoina", "CoxSnell", "Nagelkerke",
    "Tjur"),
  avg.num = 30,
  ...
)
}
\arguments{
\item{model}{A reduced or working model that needs to be investigated. The reported surrogate R-square is for this reduced model.}

\item{full_model}{A full model that needs to be specified for surrogate R-square approach.}

\item{data}{A data set contains the categorical responses and all necessary predictors (including all the predictors in the \code{full_model}).}

\item{which}{An argument to specify which pseudo R-square to calculate.}

\item{avg.num}{The number of replication for the averaging of surrogate R-square.}

\item{...}{Additional optional arguments.}
}
\value{

}
\description{
R2
}
