% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surr_rsq_ci.R
\name{surr_rsq_ci}
\alias{surr_rsq_ci}
\title{A function to calculate the interval estimate of the surrogate R-squared measure.}
\usage{
surr_rsq_ci(surr_rsq, alpha = 0.05, B = 1000, ...)
}
\arguments{
\item{surr_rsq}{A object of class \code{"surr_rsq"} that is generated by the function \code{"surr_rsq"}.
It contains the following components: \code{surr_rsq}, \code{reduced_model}, \code{full_model}, and \code{data}.}

\item{alpha}{The significance level alpha. The confidence level is 1-alpha.}

\item{B}{The number of bootstrap replications.}

\item{...}{Additional optional arguments.}
}
\value{
An list that contains the CI_lower, CI_upper, CI_lower_avg, and CI_upper_avg.
}
\description{
A function to calculate the interval estimate of the surrogate R-squared measure.
}
