% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WhiteWine}
\alias{WhiteWine}
\title{White wine}
\format{
A data frame with 4898 rows, quality score, and 11 variables of physicochemical properties of wines.
\itemize{
  \item \code{quality} Tasting preference is a rating score with ordinal values from
  0 (very bad) to 10 (excellent).

  \item \code{fixed.acidity}
  \item \code{volatile.acidity}
  \item \code{citric.acid}
  \item \code{residual.sugar}
  \item \code{chlorides}
  \item \code{free.sulfur.dioxide}
  \item \code{total.sulfur.dioxide}
  \item \code{density}
  \item \code{pH}
  \item \code{sulphates}
  \item \code{alcohol}
  }
}
\usage{
data(WhiteWine)
}
\description{
A white wine tasting preference data used in the study of Cortez, Cerdeira, Almeida, Matos,
and Reis 2009. This white wine contains 4898 white vinho verde wine samples and 12
variables including the tasting preference score of white wine and its physicochemical characteristics.
}
\examples{
head(WhiteWine)
}
\references{
Cortez, P., Cerdeira, A., Almeida, F., Matos, T., and Reis, J. (2009), “Modeling wine
preferences by data mining from physicochemical properties,” Decision Support Systems,
47, 547–553. \doi{10.1016/j.dss.2009.05.016}
}
\keyword{datasets}
