% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surr_rsq.R
\name{print}
\alias{print}
\alias{print.surr_rsq}
\title{Print surrogate R-squared measure}
\usage{
\method{print}{surr_rsq}(x, digits = max(2, getOption("digits") - 2), ...)
}
\arguments{
\item{x}{A surr_rsq object for printing out results.}

\item{digits}{A default number to specify decimal digit values.}

\item{...}{Additional optional arguments.}
}
\value{
Print surrogate R-squared measure of a surr_rsq object
}
\description{
Print surrogate R-squared measure
}
\keyword{internal}
