% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surr_rsq.R
\name{surr_rsq}
\alias{surr_rsq}
\title{surr_rsq}
\usage{
surr_rsq(model, full_model, data, avg.num = 30, ...)
}
\arguments{
\item{model}{A reduced model that needs to be investigated. The reported surrogate R-square is for this reduced model.}

\item{full_model}{A full model that contains all of the predictors in the data set.}

\item{data}{A data set contains the categorical responses, predictors.}

\item{...}{Additional optional arguments.}

\item{avg}{The number of replication for the averaging of surrogate R-square.}
}
\value{
An object of class \code{"surr_rsq"} is a list containing the following components:
\item{\code{surr_rsq}}{the surrogate R-square value;}
\item{\code{reduced_model}}{the reduced model under investigation. It should be a subset
of the full model;}
\item{\code{full_model}}{the full model used for generating the surrogate response. It should
have passed initial variable screening and model diagnostics (see Paper for reference);}
\item{\code{data}}{the dataset contains the response variable and all the predictors.}
}
\description{
surr_rsq
}
\references{
Zhu, X., Liu, D., Lin, Z. (2022). SurrRsq: an R package for evaluating goodness of fit using
surrogate R-squared
}
