% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surr_rsq_ci.R
\name{surr_rsq_ci}
\alias{surr_rsq_ci}
\title{A function to calculate the interval estimate of the surrogate R-squared measure}
\usage{
surr_rsq_ci(surr_rsq, alpha = 0.05, B = 1000, ...)
}
\arguments{
\item{surr_rsq}{A object of class \code{"surr_rsq"} that is generated by the function \code{"surr_rsq"}.
It contains the following components: \code{surr_rsq}, \code{reduced_model}, \code{full_model}, and \code{data}.}

\item{alpha}{The significance level alpha. The confidence level is 1-alpha.}

\item{B}{The number of bootstrap replications.}

\item{...}{Additional optional arguments.}
}
\value{
An object of class \code{"surr_rsq_ci"} is a list containing the following components:
\item{\code{surr_rsq}}{the surrogate R-square value;}
\item{\code{surr_rsq_ci}}{An list that contains the CI_lower, CI_upper.}
\item{\code{surr_rsq_BS}}{the surrogate R-square value from bootstrap.}
\item{\code{reduced_model}}{the reduced model under investigation. It should be a subset
of the full model;}
\item{\code{full_model}}{the full model used for generating the surrogate response. It should
have passed initial variable screening and model diagnostics (see Paper for reference);}
\item{\code{data}}{the dataset contains the response variable and all the predictors.}
}
\description{
This function generates the interval measure of surrogate R-squared by bootstrap.
}
\examples{
data("RedWine")

full_formula <- as.formula(quality ~ fixed.acidity + volatile.acidity + citric.acid
+ residual.sugar + chlorides + free.sulfur.dioxide +
total.sulfur.dioxide + density + pH + sulphates + alcohol)

fullmodel <- polr(formula = full_formula,data=RedWine, method  = "probit")

select_model <- update(fullmodel, formula. = ". ~ . - fixed.acidity -
citric.acid - residual.sugar - density")

surr_rsq_select <- surr_rsq(select_model, fullmodel, data = RedWine, avg.num = 30)

# surr_rsq_ci(surr_rsq_select, alpha = 0.05, B = 1000) # Not run, it takes time.

}
