% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surr_rsq_rank.R
\name{surr_rsq_rank}
\alias{surr_rsq_rank}
\title{The contribution of each variable in the final model}
\usage{
surr_rsq_rank(object, data, avg.num = 30, var.set = NA, ...)
}
\arguments{
\item{object}{A object of class \code{"surr_rsq"} that is generated by the function \code{"surr_rsq"}.
It contains the following components: \code{surr_rsq}, \code{reduced_model}, \code{full_model}, and \code{data}.}

\item{data}{A data set contains the categorical responses, predictors.}

\item{avg.num}{The number of replication for the averaging of surrogate R-square.}

\item{var.set}{A list that contains a few sets. Each component of these sets represents the variables
that you want to examine for the contribution of goodness of fit. Then, for one component of this list,
a model will fit by removing the specified variables.}

\item{...}{Additional optional arguments.}
}
\value{
An list that contains the contribution of Surrogate R-squared for each variable.
}
\description{
This function calculates reduction of the surrogate R-squared goodness-of-fit of each variable to measure their relative explanatory power. This function creates a table containing the reductions of surrogate R-squared by removing each one of variables in the model.
}
\examples{
data("RedWine")

full_formula <- as.formula(quality ~ fixed.acidity + volatile.acidity + citric.acid
+ residual.sugar + chlorides + free.sulfur.dioxide +total.sulfur.dioxide +
density + pH + sulphates + alcohol)

test_var_set <- list(c("volatile.acidity"),
c("chlorides"),
c("free.sulfur.dioxide"),
c("total.sulfur.dioxide"),
c("pH"),
c("sulphates"),
c("alcohol"),
c("I(sulphates^2)", "I(sulphates^3)"),
c("sulphates", "I(sulphates^2)", "I(sulphates^3)"))


fullmodel <- polr(formula = full_formula,data=RedWine, method  = "probit")
select_model <- update(fullmodel, formula. = ". ~ . - fixed.acidity -
citric.acid - residual.sugar - density")

Rank_table <- surr_rsq_rank(object  = select_model,
data    = RedWine,
var.set = test_var_set,avg.num = 30)

Rank_table$Ranking

}
