% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surr_rsq_rank.R
\name{surr_rsq_rank}
\alias{surr_rsq_rank}
\title{surr_rsq_rank}
\usage{
surr_rsq_rank(object, data, avg.num = 30, ...)
}
\arguments{
\item{object}{A object of class \code{"surr_rsq"} that is generated by the function \code{"surr_rsq"}.
It contains the following components: \code{surr_rsq}, \code{reduced_model}, \code{full_model}, and \code{data}.}

\item{data}{a data set contains the categorical responses, predictors.}

\item{...}{}

\item{avg}{the number of replication for the averaging of surrogate R-square.}
}
\description{
surr_rsq_rank
}
